Addressed PR comments :)
