+++ package.json
@@ -16,6 +16,5 @@
   },
   "workspaces": [
     "packages/*"
-  ],
-  "version": "0.2.3"
+  ]
 }
