+++ packages/widgets-manager/jest.config.js
