+++ packages/widgets-manager/jest.setup.js
