+++ packages/widgets-manager/src/constants.ts
